modules/audioclassifier/requirements.txt
azure-iot-device
azure-storage-blob
numpy
librosa
matplotlib